Refactor: split InstrumentGraph into per-instrument modules
`src/audio/InstrumentGraph.ts` had grown to 645 lines holding the sampler,
synth, and drum-machine implementations plus their shared asset-subscription
plumbing. Split each instrument into its own module under
`src/audio/instruments/`, leaving `InstrumentGraph.ts` as the `kind` dispatcher
and the published surface.

Pure refactor: no behavior change and no renamed exports. The module still
exports exactly `InstrumentNode`, `InstrumentGraphContext`,
`InstrumentNodeFactory`, `createInstrumentNode`, and `playOneShot`, so all
three consumers (`TrackAudioGraph.ts`, `ProjectAudioGraph.ts`, and
`TrackAudioGraph.test.ts`) and the 822-line `InstrumentGraph.test.ts` are
unchanged. The per-instrument factories stay private to the directory.

`instruments/types.ts` exists so the per-instrument modules can take
`InstrumentNode`/`InstrumentGraphContext` without importing the barrel that
imports them back.
