+++ CLAUDE.md
@@ -44,6 +44,12 @@ src/
 │   ├── MasterAudioGraph.ts  # The master bus's device chain and volume stage
 │   ├── DeviceChain.ts       # Ordered, ID-keyed insert-chain reconciliation shared by tracks/returns/master
 │   ├── InstrumentGraph.ts   # Sampler/synth/drum-machine instrument node factory and reconciliation
+│   ├── instruments/         # The per-instrument implementations behind `InstrumentGraph.ts`
+│   │   ├── types.ts             # `InstrumentNode`/`InstrumentGraphContext`/`InstrumentNodeFactory`, smoothing window, kind aliases
+│   │   ├── assetVoice.ts        # Reattachable per-asset buffer subscription, `playOneShot`, shared gain/pitch helpers
+│   │   ├── sampler.ts           # The sampler node: pitched, windowed voices under a per-trigger amp envelope
+│   │   ├── synth.ts             # The polyphonic subtractive synth node: PolySynth into one smoothed low-pass filter
+│   │   └── drumMachine.ts       # The drum-machine node: per-pad strips, mute/solo, choke groups, short-lived hits
 │   ├── AudioBufferCache.ts  # Asset buffer cache keyed by ID/revision with stale-load cancellation
 │   ├── toneBufferLoader.ts  # The only Tone-touching asset decode path `AudioBufferCache` uses in production
 │   ├── Transport.ts         # Play/pause/stop/seek, playhead, tempo mirror, bar loop, metronome (PRD AUD-01/AUD-02)
@@ -295,7 +301,7 @@ See [`docs/testing.md`](./docs/testing.md) for what each suite covers, how CI ga
 - `AudioRuntime` is the single application-scoped owner of the real-time Tone/Web Audio context, transport, buffer cache, and resource registry (PRD AUD-07, AUD-09; section 9.7). It is the only place production code may create, install, resume, suspend, replace, or close that context — obtain it via `getAudioRuntime()`, never construct one directly.
 - `ProjectAudioGraph` reconciles a read-only `AudioSongProjection` (from `src/projection/audioProjection.ts`) into a stable graph keyed by track, instrument, device, return, and asset IDs (PRD AUD-03, AUD-08). Passing back the exact projection object the audio projection handed out previously is a complete no-op; an edit to one track, return, or placement only touches that entity's own subgraph.
 - `TrackAudioGraph`/`ReturnAudioGraph`/`MasterAudioGraph` each own one channel strip (`Tone.PanVol`/`Tone.Volume`) plus a `DeviceChain`. `DeviceChain` reconciles an ordered `Device[]` by id: only added/removed devices create or dispose a node, and reordering relinks connections without recreating anything. Schema v1 has no concrete processors yet (Alpha Milestone 1 authors them); an unregistered `device.type` gets an inert passthrough node so topology is provable ahead of real DSP.
-- `InstrumentGraph.ts` builds the sampler/synth/drum-machine node for a track's `Instrument`. A track only replaces its instrument node when `kind` changes; an asset swap, a drum-pad added/removed, or a generic parameter edit calls the existing node's `update()` instead.
+- `InstrumentGraph.ts` builds the sampler/synth/drum-machine node for a track's `Instrument`. A track only replaces its instrument node when `kind` changes; an asset swap, a drum-pad added/removed, or a generic parameter edit calls the existing node's `update()` instead. The module itself is just the `kind` dispatcher and the public surface (`InstrumentNode`, `InstrumentGraphContext`, `InstrumentNodeFactory`, `createInstrumentNode`, `playOneShot`) — consumers import it and nothing else. Each instrument's implementation lives in its own module under `src/audio/instruments/`, over the shared types in `instruments/types.ts` and the asset-subscription helpers in `instruments/assetVoice.ts`.
 - `AudioBufferCache` decodes and caches asset buffers keyed by asset ID and content fingerprint, with reference-counted eviction and generation-tracked cancellation so a stale decode can never reconnect or overwrite a newer one. It never imports Tone itself — `toneBufferLoader.ts` is the one production loader that does, which keeps the cache's generation/refcount bookkeeping testable without any Web Audio globals.
 - `Transport.ts` owns the session-scoped playhead: play/pause/stop/seek/continue, the 40-240 BPM clamp, a bar-aligned loop range, and the metronome (PRD AUD-01/AUD-02). It is written against an injectable `TransportEngine` rather than `Tone.getTransport()`, and it never schedules the arrangement — a tempo change is mirrored onto the transport, not applied by rebuilding nodes. Tempo lives in the song and is written only by a `parameter.set` command; this layer mirrors it. `underrun.ts` counts late dispatches and reports them *sampled* as `audio_underrun`; it compares an event's intended time against `ProjectAudioGraph.audioClockNow()` (the context's true `currentTime`), never `Tone.now()`, which is `currentTime + lookAhead` and would score healthy playback as a continuous stream of drops.
 - `audioLoopPlayer.ts` plays one scheduled audio-loop event. A loop declares the tempo it was authored at, and it follows the song tempo by *time-stretching*, not resampling: `Tone.GrainPlayer` walks the buffer at `tempo / sourceTempo` while each grain sounds at its native rate, so the loop's pitch stays where it was recorded (PRD INS-02). At an unstretched rate of 1 it falls back to a plain `Tone.Player`, so a loop at its own tempo pays no granular cost at all. `audioLoopOffsetSeconds` is a position in the *buffer's* timeline and `audioLoopDurationSeconds` a span of the *song's* — the two only coincide at rate 1, and mixing them up is what silently truncates or overruns a stretched loop.
